Add timer choice and handle more cases
The host user can choose whether to have a timer on one turn,
and set the length of timer. If one player failed to give response
in limited time, the other team wins.

Handle case that user-input clue is not in the vocabulary (when
the next turn is AI spy).