redo sidebar CSS, add mobile compatibility
The arrow on the sidebar now works properly
and moves with the sidebar because it is contained
within the it.

On mobile, the sidebar now slides all the way across as
expected.

The CSS overall has also been shortened and the HTML
is more structured.

Still to do: move the board along with the sidebar. This
part had to break for the rest to work and needs reworking
too. It would probably be easier to remove "position:absolute;"
for other elements so they would flow with the sidebar.