Add javascript testing framework
The testing framework is MochaJS with ChaiJS.
The framework needs node package manager, which can be
installed with `apt install npm` or equivalent.

Run `npm install mocha chai --save-dev` in the /test/js/
directory to get the framework.
In /test/js/test/ run `npm run test` to run the test files in
that directory.

Tests should soon be implemented on files like board.js and
others.