JS test framework made browser compatible.
The instructions to test are largely the same as
the dependencies needed are still the same (npm, mocha and chai).

The change is that the tests results are now viewed in
/test/js/runner.html and the test scripts are written
in /test/js/test.js.

This change was made because browser modules were not
working with the mocha terminal.