Implement game difficulty in single player mode
Game difficulty is controlled by AI spy in all-AI team.
The easier game, the more mistake AI spy will make.
(ONLY in single player mode)

Note that the AI spy in player's team will not be affected and maintains
highest accuracy as well as in multi player mode.

Merge changes in GameDev to prepare for timer bar in next step.