add working python testing after many hours
Tests can be added in test.py. To write a test follow these steps:
1) write `client_.emit(<protocol>, <message>)` to send a message
to the server.
2) write a `client_on(<protocol>)` statement to receive a message.
3) handle the recieved message by declaring a function with
one parameter (the data received).
4) write an assert statement in that function.
(See the Template Test for an example to follow).

To run the test script run `./runTest.sh` from the test/python
directory OR run the server and the scripts manually with
python3.

This is a bit scuffed and doesn't follow a good framework,
but after many hours this is the best I could do.